# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.ArchiveListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handlers.ArchiveResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  handlers.ArchiveResponse:
    properties:
      artists:
        items:
          type: string
        type: array
      category:
        type: string
      characters:
        items:
          type: string
        type: array
      circles:
        items:
          type: string
        type: array
      created_at:
        type: string
      file_path:
        type: string
      id:
        type: string
      is_favorited:
        type: boolean
      language:
        type: string
      page_count:
        type: integer
      parodies:
        items:
          type: string
        type: array
      progress:
        $ref: '#/definitions/handlers.ProgressResponse'
      rating:
        type: integer
      release_date:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      thumbs_ready:
        type: boolean
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.ProgressResponse:
    properties:
      completed:
        type: boolean
      current_page:
        type: integer
      last_read:
        type: string
    type: object
  handlers.ReaderSettingsResponse:
    properties:
      background:
        type: string
      fit_mode:
        type: string
      page_layout:
        type: string
      reading_direction:
        type: string
      view_mode:
        type: string
    type: object
...
  handlers.TagListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handlers.TagResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  handlers.TagResponse:
    properties:
      count:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  handlers.UpdateReaderSettingsRequest:
    properties:
      background:
        type: string
      fit_mode:
        type: string
      page_layout:
        type: string
      reading_direction:
        type: string
      view_mode:
        type: string
    type: object
  handlers.UpdateTagDescriptionRequest:
    properties:
      description:
        type: string
    type: object
...
  response.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Self-hosted doujinshi library manager API
  title: Shoka API
  version: "1.0"
paths:
...
  /api/reader-settings:
    get:
      description: Returns default values if the user hasn't customized anything yet.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ReaderSettingsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get the current user's reader settings
      tags:
      - reader-settings
    patch:
      consumes:
      - application/json
      description: Partial update - omitted fields keep their current (or default)
        value. Creates the row on first call.
      parameters:
      - description: Fields to change
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateReaderSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ReaderSettingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update the current user's reader settings
      tags:
      - reader-settings
...
  /api/tags:
    get:
      parameters:
      - default: 1
        description: Page number (1-based)
        in: query
        name: page
        type: integer
      - default: 50
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TagListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: List tags with pagination
      tags:
      - tags
  /api/tags/{id}:
    delete:
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Delete a tag
      tags:
      - tags
    patch:
      consumes:
      - application/json
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: Description
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateTagDescriptionRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update a tag's description
      tags:
      - tags
  /api/tags/{name}:
    get:
      parameters:
      - description: Tag name
        in: path
        name: name
        required: true
        type: string
      - default: 1
        description: Page number (1-based)
        in: query
        name: page
        type: integer
      - default: 24
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ArchiveListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: List archives for a tag
      tags:
      - tags
  /api/tags/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.TagResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: List every tag
      tags:
      - tags
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"